+++ cartads/lib/Util.php
@@ -10,7 +10,13 @@ class Util {
     public static function projectIsCartADS(string $repo, string $projectName) {
         $project = \lizmap::getProject($repo . '~' . $projectName);
 
-        if (!$project || substr($projectName, 0, 7) !== 'cartads' || substr($projectName, -7) !== 'cartads') {
+        if (!$project) {
+            // Projet inconnu
+            return self::ERR_CODE_PROJECT_NAME;
+        }
+
+        if (substr($projectName, 0, 7) !== 'cartads' && substr($projectName, -7) !== 'cartads') {
+            // le projet ne commence ni ne fini par 'cartads'
             return self::ERR_CODE_PROJECT_NAME;
         }